подправила проект
