+++ students/Ostrovskaya Elena/Others/Shop/src/index.js
@@ -1,5 +1,6 @@
-import app from './public/js/main.js'
+import './public/js/main.js'
 import './public/style/style.css'
 import './public/style/normalize.css'
 
-app()
+
+
