+++ students/Ostrovskaya Elena/Others/Shop/src/public/js/main.js
@@ -1,0 +1,24 @@
+
+ //ИМИТАЦИЯ РАБОТЫ БАЗЫ ДАННЫХ И СЕРВЕРА
+
+ let PRODUCTS_NAMES = ['Processor', 'Display', 'Notebook', 'Mouse', 'Keyboard']
+ let PRICES = [100, 120, 1000, 15, 18]
+ let IDS = [0, 1, 2, 3, 4]
+ let IMGS = ['https://cs8.pikabu.ru/post_img/big/2017/12/25/5/1514188160141511997.jpg', 
+ 'https://store.storeimages.cdn-apple.com/8756/as-images.apple.com/is/HMUB2?wid=1144&hei=1144&fmt=jpeg&qlt=80&op_usm=0.5,0.5&.v=1563827752399',
+ 'https://zeon18.ru/files/item/Xiaomi-Mi-Notebook-Air-4G-Officially-Announced-Weboo-co-2%20(1)_1.jpg',
+ 'https://files.sandberg.it/products/images/lg/640-05_lg.jpg',
+ 'https://images-na.ssl-images-amazon.com/images/I/81PLqxtrJ3L._SX466_.jpg']
+
+
+
+ console.log("main");
+
+ window.addEventListener('load', function () {
+   console.log("load");
+   let cart = new Cart ();
+   let catalog = new Catalog();
+
+   catalog._init(cart);
+   cart._handleEvents();
+})
